% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{bbox}
\alias{bbox}
\title{Get and show bounding box}
\usage{
bbox(west, south, east, north, plot = TRUE)
}
\arguments{
\item{west}{Minimum longitude in decimal degrees}

\item{south}{Minimum latitude in decimal degrees}

\item{east}{Maximum longitude in decimal degrees}

\item{north}{Maximum latitude in decimal degrees}

\item{plot}{Generate a plot? Requires leaflet package.}
}
\description{
Get and show bounding box
}

