% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{charNameLookup}
\alias{charNameLookup}
\title{Lookup characteristic names available from WQP}
\usage{
charNameLookup(pattern, ignore.case = TRUE, ...)
}
\arguments{
\item{pattern}{text string to match}

\item{...}{other arguments passed to grepl}
}
\description{
Lookup characteristic names available from WQP
}

