% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{getConcData}
\alias{getConcData}
\title{Improvements on dataRetrieval functions.
Gets data for rating-curve models. That is, constituent concentration and flows on river monitoring stations.
returns an object of class rcData}
\usage{
getConcData(bBox = NULL, charname = NULL, chartype = NULL,
  statecode = NULL, ...)
}
\arguments{
\item{bBox}{bounding box, can use bbox()}

\item{charname}{characteristic name. See charNameLookup}
}
\description{
Improvements on dataRetrieval functions.
Gets data for rating-curve models. That is, constituent concentration and flows on river monitoring stations.
returns an object of class rcData
}

