% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{getFlowData}
\alias{getFlowData}
\title{Simple flow retrieval from WQP sites}
\usage{
getFlowData(siteids, n = 100)
}
\arguments{
\item{siteids}{Character vector of site identifiers}

\item{n}{Maximum number of sites per retrieval. Full retrieval will be done in batches of n.}
}
\description{
Simple flow retrieval from WQP sites
}

