% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{getWQPCodes}
\alias{getWQPCodes}
\title{Retrieve data.frame of all characteristic names available from WQP}
\usage{
getWQPCodes(name = c("statecode", "countycode", "Sitetype", "Organization",
  "sampleMedia", "Characteristictype", "Characteristicname", "providers"))
}
\description{
Retrieve data.frame of all characteristic names available from WQP
}

