% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{statecodeLookup}
\alias{statecodeLookup}
\title{Lookup a state via the FIPS state code}
\usage{
statecodeLookup(code = NULL, abb = NULL)
}
\arguments{
\item{code}{FIPS state code}

\item{abb}{State abbreviation (i.e. postal code)}
}
\description{
Lookup a state via the FIPS state code
}

