% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timeit}
\alias{timeit}
\title{Simple timing of function execution}
\usage{
timeit(expr)
}
\arguments{
\item{expr}{Unevaluated function call.}
}
\description{
Time the execution of a function, returning the function's output and messaging the time difference
}

