% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toUTC_oneTZ}
\alias{toUTC}
\alias{toUTC_oneTZ}
\title{Convert timezones of a datetime character vector}
\usage{
toUTC_oneTZ(timestring, tzstring, ...)

toUTC(timestring, tzstring, ...)
}
\arguments{
\item{timestring}{A character vector convertible to POSIXct format}

\item{tzstring}{A character vector specifying time zones of timestring.
For toUTC_oneTZ, this must be of length one. For toUTC, this must have length
equal to that of timestring.}

\item{...}{Arguments passed to as.POSIXct. (e.g. \code{format})}
}
\description{
Conversion for a single timezone
}
\details{
toUTC_oneTZ allows conversion for observations in a single timezone to UTC,
while toUTC allows for multiple starting timezones to be specified.
}
\section{Functions}{
\itemize{
\item \code{toUTC}: Conversion allowing for multiple timezones
}}
\seealso{
\code{\link{timezones}} for how to specify timezones, \code{\link{as.POSIXct}}
}

