% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{validateUnits}
\alias{validateUnits}
\title{replacement function for commonly given, but "improper" units,
e.g. "CFS", which *should* be ft3/s}
\usage{
validateUnits(unit)
}
\description{
replacement function for commonly given, but "improper" units,
e.g. "CFS", which *should* be ft3/s
}

