% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_checkActivity}
\alias{wqp_checkActivity}
\title{Check activity types}
\usage{
wqp_checkActivity(wqpData)
}
\arguments{
\item{wqpData}{data.frame returned by readWQPData}
}
\description{
Check activity types
}

