% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_checkBDL}
\alias{wqp_checkBDL}
\title{Check for detection limit behavior}
\usage{
wqp_checkBDL(wqpData)
}
\arguments{
\item{wqpData}{data.frame returned by readWQPData}
}
\description{
This function performs the following:
1. Adds is.bdl column
2. Converts detection-limit units to MeasureValue units.
3. Augments DetectionQuantitationLimitMeasure.MeasureValue,
DetectionQuantitationLimitMeasure.MeasureUnitCode.
4. If detection limit not reported, assumes it is at largest non-bdl value for that dataset
}

