% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_checkClasses}
\alias{wqp_checkClasses}
\title{Make sure all columns are present and classes are correct.}
\usage{
wqp_checkClasses(wqpData)
}
\arguments{
\item{wqpData}{data.frame returned by readWQPData}
}
\description{
Make sure all columns are present and classes are correct.
}

