% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_checkDepth}
\alias{wqp_checkDepth}
\title{Check depth measurements}
\usage{
wqp_checkDepth(wqpData, units = "m")
}
\arguments{
\item{wqpData}{data.frame returned by readWQPData}

\item{units}{Desired units for depth measurements}
}
\description{
Assures that depth units are same and reference text is familiar.
}

