% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_checkDuplicates}
\alias{wqp_checkDuplicates}
\title{Check for and remove duplicates in a WQP data.frame}
\usage{
wqp_checkDuplicates(wqpData)
}
\arguments{
\item{wqpData}{data.frame returned by readWQPData}
}
\description{
unlike \code{unique}, this preserves attributes and a record of the check.
}

