% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_checkFraction}
\alias{wqp_checkFraction}
\title{Filters wqpData to include only allowed fractions}
\usage{
wqp_checkFraction(wqpData, silent = FALSE)
}
\arguments{
\item{wqpData}{data.frame returned by readWQPData}

\item{silent}{If FALSE (default) report how many rows and which fractions were omitted.}
}
\description{
Filters wqpData to include only allowed fractions
}

