% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_checkTZ}
\alias{wqp_checkTZ}
\title{Check timezones for sites in wqpData}
\usage{
wqp_checkTZ(wqpData, geonamesUser = "markwh")
}
\arguments{
\item{wqpData}{data.frame returned by readWQPData}
}
\description{
Uses geonames API, for which you need a key. Uses mine for now.
}

