% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_complicateConc}
\alias{wqp_complicateConc}
\title{Unsimplify wqp concentration data}
\usage{
wqp_complicateConc(simpleConc)
}
\arguments{
\item{wqpData}{processed WQP data result}
}
\description{
Unsimplify wqp concentration data
}

