% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_setAttrs}
\alias{wqp_setAttrs}
\title{Sets attributes of object to attributes only when such attributes aren't present in object already.}
\usage{
wqp_setAttrs(object, attributes, check = NULL)
}
\description{
Sets attributes of object to attributes only when such attributes aren't present in object already.
}

