% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_simplifyFlow}
\alias{wqp_simplifyFlow}
\title{Simplify wqp flow data}
\usage{
wqp_simplifyFlow(wqpData, average = c("none", "time"),
  allowDuplicates = FALSE)
}
\arguments{
\item{wqpData}{processed WQP data result}

\item{allowDuplicates}{If \code{FALSE}, remove duplicates using \code{unique}
following simplification.}
}
\description{
Simplify wqp flow data
}

