% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqpToRcData}
\alias{wqpToRcData}
\title{Takes an object returned by readWQPData (or getConcData) and returns an
object of class rcData}
\usage{
wqpToRcData(concdata, flowdata)
}
\description{
Takes an object returned by readWQPData (or getConcData) and returns an
object of class rcData
}

